chore: restructuring, move crypto functions to ./src/crypto
