+++ src/crypto/aes.cc
@@ -38,9 +38,11 @@ NOTE:   String length must be evenly divisible by 16byte (str_len % 16 == 0)
 /*****************************************************************************/
 /* Includes:                                                                 */
 /*****************************************************************************/
+#include "crypto/aes.h"
+
 #include <stdint.h>
 #include <string.h> // CBC mode, for memset
-#include "aes.h"
+
 
 /*****************************************************************************/
 /* Defines:                                                                  */
