+++ src/crypto/aes.h
@@ -1,8 +1,8 @@
 // Source from https://github.com/kokke/tiny-AES128-C
 // Public Domain / CC0 / Unlicense
 // Copyright (C) 2017-2019 kokke (CC0-1.0)
-#ifndef _AES_H_
-#define _AES_H_
+#ifndef CRYPTO_AES_H_
+#define CRYPTO_AES_H_
 
 #include <stdint.h>
 
@@ -41,4 +41,4 @@ void AES_CBC_decrypt_buffer(uint8_t* output, uint8_t* input, uint32_t length, co
 #endif // #if defined(CBC) && (CBC == 1)
 
 
-#endif //_AES_H_
+#endif // CRYPTO_AES_H_
