+++ src/crypto/sha256.cc
@@ -17,7 +17,7 @@
 //  IMPORTS
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-#include "sha256.h"
+#include "crypto/sha256.h"
 #include <memory.h>
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
